requirements.txt
accelerate==1.8.1
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
blis==1.3.0
catalogue==2.0.10
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpathlib==0.21.1
confection==0.1.5
cryptography==45.0.5
cymem==2.0.11
Deprecated==1.2.18
easyocr==1.7.2
ecdsa==0.19.1
faiss-cpu==1.11.0
fastapi==0.115.14
filelock==3.18.0
fsspec==2025.5.1
h11==0.16.0
hf-xet==1.1.5
hiredis==3.2.1
huggingface-hub==0.33.2
idna==3.10
imageio==2.37.0
Jinja2==3.1.6
joblib==1.5.1
langcodes==3.5.0
language_data==1.3.0
lazy_loader==0.4
limits==5.4.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.13
networkx==3.5
ninja==1.11.1.4
numpy==2.3.1
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
opencv-python-headless==4.11.0.86
packaging==25.0
passlib==1.7.4
pillow==11.3.0
preshed==3.0.10
psutil==7.0.0
pyasn1==0.6.1
pyclipper==1.3.0.post6
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyMuPDF==1.26.3
python-bidi==0.6.6
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.2
redis==6.2.0
regex==2024.11.6
requests==2.32.4
rich==14.0.0
rsa==4.9.1
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.7.0
scipy==1.16.0
sentence-transformers==5.0.0
setuptools==80.9.0
shapely==2.1.1
shellingham==1.5.4
six==1.17.0
slowapi==0.1.9
smart_open==7.3.0.post1
sniffio==1.3.1
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
starlette==0.46.2
sympy==1.14.0
thinc==8.3.6
threadpoolctl==3.6.0
tifffile==2025.6.11
tokenizers==0.21.2
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
transformers==4.53.0
triton==3.3.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
uvicorn==0.35.0
wasabi==1.1.3
weasel==0.4.1
wrapt==1.17.2
